Deleting other user's friends is forbidden

+++ backend/profiles/views.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.models import User
 from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from rest_framework import status
@@ -6,7 +7,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from .models import (FriendRequest, Friendship, MatchlistDislike,
-                     MatchlistLike, Profile, User, WatchlistMovie)
+                     MatchlistLike, Profile, WatchlistMovie)
 from .serializers import (FriendRequestSerializer, FriendshipSerializer,
                           MatchlistDislikeSerializer, MatchlistLikeSerializer,
                           ProfileSerializer, UserSerializer,
@@ -101,13 +102,16 @@ class FriendshipAPIView(APIView):
 
     permission_classes = (IsAuthenticated,)
 
-    def get(self, request):
-        friends = Friendship.objects.filter(Q(user=request.user) | Q(friend=request.user))
-        serializer = FriendshipSerializer(friends, many=True, context={'user_id': request.user.id})
+    def get(self, request, user_id=-1):
+        user = request.user if user_id < 0 else get_object_or_404(User, id=user_id)
+        friends = Friendship.objects.filter(Q(user=user) | Q(friend=user))
+        serializer = FriendshipSerializer(friends, many=True, context={'user_id': user.id})
         return Response(data=serializer.data)
 
     def delete(self, request):
         friendship = get_object_or_404(Friendship, id=request.data['id'])
+        if friendship.user != request.user and friendship.friend != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         friendship.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
